+++ src/main/kotlin/com/zepben/ewb/database/sql/common/BaseDatabaseWriter.kt
@@ -22,7 +22,7 @@ import java.sql.SQLException
  * @property databaseTables The tables to create in the database.
  * @property databaseInitialiser The hooks used to initilise the database.
  */
-abstract class BaseDatabaseWriter<TTables : BaseDatabaseTables, T> internal constructor() {
+abstract class BaseDatabaseWriter<TTables : BaseDatabaseTables> internal constructor() {
 
     protected val logger: Logger = LoggerFactory.getLogger(javaClass)
 
@@ -32,11 +32,11 @@ abstract class BaseDatabaseWriter<TTables : BaseDatabaseTables, T> internal cons
     /**
      * Write to the database.
      *
-     * @param data The data to write to the database.
+     * @param writeData Function to perform writes to the database
      *
      * @return true if the database was successfully written, otherwise false.
      */
-    fun write(data: T): Boolean {
+    fun connectAndWrite(writeData: () -> Boolean): Boolean {
         try {
             if (!databaseInitialiser.beforeConnect(logger))
                 return false
@@ -45,7 +45,7 @@ abstract class BaseDatabaseWriter<TTables : BaseDatabaseTables, T> internal cons
                 databaseInitialiser.afterConnectBeforePrepare(connection, databaseTables, logger)
                     && versionMatches(connection)
                     && prepareInsertStatements(connection)
-                    && writeData(data)
+                    && writeData()
                     && databaseInitialiser.afterWriteBeforeCommit(connection, databaseTables, logger)
                     && commit(connection)
             }
@@ -60,14 +60,6 @@ abstract class BaseDatabaseWriter<TTables : BaseDatabaseTables, T> internal cons
             }
         }
     }
-
-    /**
-     * Write the actual data to the database.
-     *
-     * @param data The data to write.
-     * @return true if the processing was successful, otherwise false.
-     */
-    protected abstract fun writeData(data: T): Boolean
 
     private fun connect(): Connection {
         logger.info("Connecting to database...")